requirements.txt
click==7.1.2
ffmpeg-python==0.2.0
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
